Separate developement code from production code
This commit moves all the assets to the root folder. After that, I add
build command on the `package.json` to create build process when it gets
deployed to Netlify. This way, the history of the line can be
read—minified assets are impossible to track.

This commit also contains UI changes to the "Top keywords" section.
Also, change from "Top 10 keywords" to "Top keywords". So, the
application is showing all the keywords instead of just the top ten
keywords.

This commit creates more statistics by counting characters without
spaces and lines of text.